Move mask related functions.
