+++ depthai_nodes/ml/parsers/utils/masks_utils.py
@@ -1,3 +1,6 @@
+from typing import List, Optional, Tuple
+
+import depthai as dai
 import numpy as np
 
 
@@ -61,3 +64,34 @@ def process_single_mask(
     mask = sigmoid(np.sum(protos * mask_coeff[..., np.newaxis, np.newaxis], axis=0))
     mask = crop_mask(mask, scaled_bbox)
     return (mask > mask_conf).astype(np.uint8)
+
+
+def get_segmentation_outputs(
+    output: dai.NNData,
+    mask_output_layer_names: Optional[List[str]] = None,
+    protos_output_layer_name: Optional[str] = None,
+) -> Tuple[List[np.ndarray], np.ndarray, int]:
+    """Get the segmentation outputs from the Neural Network data."""
+    # Get all the layer names
+    layer_names = mask_output_layer_names or output.getAllLayerNames()
+    mask_outputs = sorted([name for name in layer_names if "mask" in name])
+    masks_outputs_values = [
+        output.getTensor(o, dequantize=True).astype(np.float32) for o in mask_outputs
+    ]
+    protos_output = output.getTensor(
+        protos_output_layer_name or "protos_output", dequantize=True
+    ).astype(np.float32)
+    protos_len = protos_output.shape[1]
+    return masks_outputs_values, protos_output, protos_len
+
+
+def reshape_seg_outputs(
+    protos_output: np.ndarray,
+    protos_len: int,
+    masks_outputs_values: List[np.ndarray],
+) -> Tuple[np.ndarray, int, List[np.ndarray]]:
+    """Reshape the segmentation outputs."""
+    protos_output = protos_output.transpose((0, 3, 1, 2))
+    protos_len = protos_output.shape[1]
+    masks_outputs_values = [o.transpose((0, 3, 1, 2)) for o in masks_outputs_values]
+    return protos_output, protos_len, masks_outputs_values
